refactor(legacy): extract @nodejs/doc-kit-generator-legacy package
