+++ packages/legacy/src/legacy-html-all/generate.mjs
@@ -3,9 +3,10 @@
 import { readFile, writeFile } from 'node:fs/promises';
 import { join } from 'node:path';
 
-import getConfig from '../../utils/configuration/index.mjs';
-import { minifyHTML } from '../../utils/html-minifier.mjs';
-import { getRemarkRehype as remark } from '../../utils/remark.mjs';
+import getConfig from '@node-core/doc-kit/utils/configuration/index.mjs';
+import { minifyHTML } from '@node-core/doc-kit/utils/html-minifier.mjs';
+import { getRemarkRehype as remark } from '@node-core/doc-kit/utils/remark.mjs';
+
 import { replaceTemplateValues } from '../legacy-html/utils/replaceTemplateValues.mjs';
 import tableOfContents from '../legacy-html/utils/tableOfContents.mjs';
