+++ packages/legacy/src/legacy-html/utils/__tests__/slugger.test.mjs
