+++ packages/legacy/src/legacy-json-all/index.mjs
@@ -14,7 +14,7 @@ export default {
   description:
     'Generates the `all.json` file from the `legacy-json` generator, which includes all the modules in one single file.',
 
-  dependsOn: '@node-core/doc-kit/legacy-json',
+  dependsOn: '@nodejs/doc-kit-generator-legacy/legacy-json',
 
   defaultConfiguration: {
     minify: false,
