+++ packages/legacy/src/legacy-json/constants.mjs
