+++ packages/legacy/src/legacy-json/utils/__tests__/buildHierarchy.test.mjs
