+++ packages/legacy/src/legacy-json/utils/__tests__/buildSection.test.mjs
